refactored to dedicated olson_2024_spike_gadgets_recording_interface
