+++ src/jadhav_lab_to_nwb/olson_2024/__init__.py
@@ -1,2 +1,3 @@
-from .olson_2024_behaviorinterface import Olson2024BehaviorInterface
+from .olson_2024_behavior_interface import Olson2024BehaviorInterface
+from .olson_2024_spike_gadgets_recording_interface import Olson2024SpikeGadgetsRecordingInterface
 from .olson_2024_nwbconverter import Olson2024NWBConverter
