+++ src/jadhav_lab_to_nwb/olson_2024/olson_2024_convert_session.py
@@ -3,11 +3,9 @@ from pathlib import Path
 import datetime
 from zoneinfo import ZoneInfo
 import shutil
-from xml.etree import ElementTree
 
 from neuroconv.utils import load_dict_from_file, dict_deep_update
 from neuroconv.datainterfaces import SpikeGadgetsRecordingInterface
-from neuroconv import ConverterPipe
 
 from jadhav_lab_to_nwb.olson_2024 import Olson2024NWBConverter
 
@@ -23,28 +21,15 @@ def session_to_nwb(data_dir_path: str | Path, output_dir_path: str | Path, stub_
     session_id = "sample_session"
     nwbfile_path = output_dir_path / f"{session_id}.nwb"
 
-    data_interfaces = dict()
+    source_data = dict()
     conversion_options = dict()
 
-    # Add Ephys Recording Interface
+    # Add Ephys
     file_path = data_dir_path / f"{data_dir_path.name}.rec"
-    RecordingInterface = SpikeGadgetsRecordingInterface(file_path=file_path)
-    recording_extractor = RecordingInterface.recording_extractor
-    channel_ids = recording_extractor.get_channel_ids()
-    channel_names = recording_extractor.get_property(key="channel_name", ids=channel_ids)
-    header = get_spikegadgets_header(file_path)
-    hwChan_to_nTrode = extract_hwChan_to_nTrode(header)
-    group_names = []
-    for channel_name in channel_names:
-        hwChan = channel_name.split("hwChan")[-1]
-        nTrode = hwChan_to_nTrode[hwChan]
-        group_names.append(f"nTrode{nTrode}")
-    recording_extractor.set_property(key="group_name", ids=channel_ids, values=group_names)
-
-    data_interfaces.update(dict(Recording=RecordingInterface))
+    source_data.update(dict(Recording=dict(file_path=file_path)))
     conversion_options.update(dict(Recording=dict(stub_test=stub_test)))
 
-    converter = ConverterPipe(data_interfaces=data_interfaces)
+    converter = Olson2024NWBConverter(source_data=source_data)
 
     # Add datetime to conversion
     metadata = converter.get_metadata()
@@ -57,65 +42,6 @@ def session_to_nwb(data_dir_path: str | Path, output_dir_path: str | Path, stub_
 
     # Run conversion
     converter.run_conversion(metadata=metadata, nwbfile_path=nwbfile_path, conversion_options=conversion_options)
-
-
-def get_spikegadgets_header(file_path: str | Path):
-    """Get the header information from a SpikeGadgets .rec file.
-
-    This function reads the .rec file until the "</Configuration>" tag to extract the header information.
-
-    Parameters
-    ----------
-    file_path : str | Path
-        Path to the .rec file.
-
-    Returns
-    -------
-    str
-        The header information from the .rec file.
-
-    Raises
-    ------
-    ValueError
-        If the header does not contain "</Configuration>".
-    """
-    header_size = None
-    with open(file_path, mode="rb") as f:
-        while True:
-            line = f.readline()
-            if b"</Configuration>" in line:
-                header_size = f.tell()
-                break
-
-        if header_size is None:
-            ValueError("SpikeGadgets: the xml header does not contain '</Configuration>'")
-
-        f.seek(0)
-        return f.read(header_size).decode("utf8")
-
-
-def extract_hwChan_to_nTrode(header_txt):
-    """Extract the hardware channel to tetrode mapping from the header text.
-
-    Parameters
-    ----------
-    header_txt : str
-        The header text.
-
-    Returns
-    -------
-    dict
-        A dictionary mapping hardware channel IDs to tetrode IDs.
-    """
-    root = ElementTree.fromstring(header_txt)
-    sconf = root.find("SpikeConfiguration")
-    hwChan_to_nTrode = dict()
-    for tetrode in sconf.findall("SpikeNTrode"):
-        nTrode = tetrode.attrib["id"]
-        for electrode in tetrode.findall("SpikeChannel"):
-            hwChan = electrode.attrib["hwChan"]
-            hwChan_to_nTrode[hwChan] = nTrode
-    return hwChan_to_nTrode
 
 
 if __name__ == "__main__":
