+++ src/jadhav_lab_to_nwb/olson_2024/olson_2024_nwbconverter.py
@@ -1,15 +1,12 @@
 """Primary NWBConverter class for this dataset."""
 from neuroconv import NWBConverter
-from neuroconv.datainterfaces import (
-    SpikeGadgetsRecordingInterface,
-)
 
-from jadhav_lab_to_nwb.olson_2024 import Olson2024BehaviorInterface
+from jadhav_lab_to_nwb.olson_2024 import Olson2024BehaviorInterface, Olson2024SpikeGadgetsRecordingInterface
 
 
 class Olson2024NWBConverter(NWBConverter):
     """Primary conversion class for my extracellular electrophysiology dataset."""
 
     data_interface_classes = dict(
-        Recording=SpikeGadgetsRecordingInterface,
+        Recording=Olson2024SpikeGadgetsRecordingInterface,
     )
